-项目结构优化 -修复输出包名，无法卸载的问题
